Add files via upload
This code inputs the Bangla characters as strokes( The dataset is co-ordinates), clusters them into 64 clusters. The results are put in an HDF file for easy handling of data. The code also renders the image in each group in the HDF, so that, one can manually see the clustering results. The code, though works correctly, discards more than 50 percent of the training data, due to divide by zero issues.